Cells are actors too!
The term "actor" is a bit overloaded. Celluloid really provides
concurrent objects which are implemented as an asbtraction on top of
actors, so while they are actors, having a separate term seems
appropriate.

I'm a big fan of "cell" for this term and have heard other people start
using it organically even before I began to suggest it myself.