Lower shutdown timeout to 10 seconds and print warning
Many people have complained about being unable to terminate Celluloid
programs with ^C and so forth when Celluloid is not able to cleanly
shut down all actors.

Celluloid has a built-in failsafe for this, but it does not kick in
until after 120 seconds! Lowering this failsafe time should make it
easier for people who are having trouble shutting down Celluloid-based
programs at the risk that there are reasonable circumstances under which
10 seconds isn't enough.