change actor use a TaskSet. fixes #157
On JRuby the runtime will raise an exception when a collection
is simultaneously being iterated and modified. This happens when
the "actors.tasks" collection is being iterated during a stack
dump while task_fiber is also adding self to actor.tasks. By
making the collection more "functional" we avoid this error.